requirements.txt
zeep==3.4.0
Django==2.2.1
djangorestframework==3.9.4
